+++ HISTORY.md
@@ -7,6 +7,8 @@ Work in progress
 - yalafi.shell
   - better error handling if position mapping fails for text report
     (issue [#75](../../issues/75))
+- added tests for yalafi.shell, using option --lt-command and temporarily
+  created LT emulations
 - Vim script vlty.vim (for vimtex): minor update
 - README.md: minor edits